requirement.txt
requests
scapy
python3-nmap
paramiko
beautifulsoup4
hashlib